Proyecto-Anatomia-medac/backend/requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
django-filter
mysqlclient
Pillow